Improve Bluetooth backend (incomplete)
Devices can now connect and send messages, but there
are still a sizable number of bugs. For one, the
messages aren't actually reconstructed into
NetworkPackages.

The current implementation enforces a synchronized
connection model which may need to be adjusted to
use e.g. queues.

Also, I think I might be able to avoid connecting
to devices that lack a KDE Connect bluetooth service.

It would require a blacklist in SharedPreferences for
 outgoing identity packets.